Merge: Lazy Loading for Executable (#18)
* Fix: kernel panic

* Feat: setup_stack

setup_stack() 구현
- 스택용 페이지 할당
- rsp 설정
vm.c
- 포맷팅
- 주석 추가

* Merge: Lazy Loading for Executable

---------

+++ pintos/include/vm/vm.h
@@ -122,7 +122,7 @@ void vm_dealloc_page(struct page *page);
 bool vm_claim_page(void *va);
 enum vm_type page_get_type(struct page *page);
 
-unsigned page_hash_func(const struct hash_elem *elem, void *aux UNUSED);
+uint64_t page_hash_func(const struct hash_elem *elem, void *aux UNUSED);
 bool compare_hash_adrr(const struct hash_elem *a, const struct hash_elem *b,
                        void *aux UNUSED);
